% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapTrend.R
\name{mapTrend}
\alias{mapTrend}
\title{Convert a series of raster files to a single phenology raster.}
\usage{
mapTrend(File_List, Year_List, parallel = FALSE, n.cores = NULL,
  verbose = FALSE)
}
\arguments{
\item{File_List}{List of phenology raster files (i.e. those produced in `mapPheno`)}

\item{Year_List}{Integer Year (YYYY)}

\item{parallel}{TRUE or FALSE (Default = FALSE) if TRUE, use parallel backend through plyr::aaply}

\item{n.cores}{Integer number of cores to be used for parallel processing (only use if parallel = TRUE)}

\item{verbose}{TRUE or FALSE (Default = FALSE)}
}
\value{
Raster object with extent=extent(raster(File_List)[1]) and CRS = crs(raster(File_List)[1]).  Digital numbers are expressed as Day of Year.
}
\description{
Convert a series of raster files to a single phenology raster.
}
\examples{
\dontshow{
fpath <- system.file("extdata", package="phenomap")
File_List <- paste(fpath, list.files(path = fpath, pattern=c("ExCrop_")), sep="/")
File_List <- File_List[1:2]

PhenoFactor = "VI"
phase = "greenup"
threshold = 0.5
year = 2016
NDVI = 1
VIQ = 3
DOY = 4
PR = 5
parallel = FALSE
n.cores = NA
verbose = TRUE

Sample.Greenup <- mapPheno(File_List = File_List, PhenoFactor = PhenoFactor,
                           phase = phase, threshold = threshold, year = year,
                           NDVI = NDVI, VIQ = VIQ, DOY = DOY, PR = PR,
                           SnowExtent=SnowExtent,
                           parallel = parallel, n.cores = n.cores,
                           verbose = verbose)
}
\dontrun{

fpath <- system.file("extdata", package="phenomap")
File_List <- paste(fpath, list.files(path = fpath, pattern=c("TinyCrop_")), sep="/")
File_List

PhenoFactor = "VI"
phase = "greenup"
threshold = 0.5
year = 2016
NDVI = 1
VIQ = 3
DOY = 4
PR = 5
parallel = FALSE
n.cores = NA
verbose = TRUE

Sample.Greenup <- mapPheno(File_List = File_List, PhenoFactor = PhenoFactor,
                           phase = phase, threshold = threshold, year = year,
                           NDVI = NDVI, VIQ = VIQ, DOY = DOY, PR = PR,
                           SnowExtent = SnowExtent,
                           parallel = parallel, n.cores = n.cores,
                           verbose = verbose)

}
}
